in update_slave_conf, first update the script, exec and then update the rock admin scripts

+++ update_slave_conf.sh
@@ -1,8 +1,13 @@
 #! /bin/sh -ex
 
+if test -z "DONE_SLAVE_CONF"; then
+    cd /home/build/slave_conf
+    git remote update
+    git reset --hard origin/master
+    export DONE_SLAVE_CONF=1
+    exec $0
+fi
+
 cd /home/build/rock_admin_scripts
-git remote update
-git reset --hard origin/master
-cd /home/build/slave_conf
 git remote update
 git reset --hard origin/master
